+++ cluster-autoscaler/core/utils/utils_test.go
@@ -21,7 +21,8 @@ import (
 	"time"
 
 	"k8s.io/autoscaler/cluster-autoscaler/utils/taints"
-	. "k8s.io/autoscaler/cluster-autoscaler/utils/test"
+	no "k8s.io/autoscaler/cluster-autoscaler/utils/test/node"
+	po "k8s.io/autoscaler/cluster-autoscaler/utils/test/pod"
 
 	"github.com/stretchr/testify/assert"
 	apiv1 "k8s.io/api/core/v1"
@@ -29,11 +30,11 @@ import (
 )
 
 func TestSanitizePods(t *testing.T) {
-	pod := BuildTestPod("p1", 80, 0)
+	pod := po.BuildTestPod("p1", 80, 0)
 	pod.Spec.NodeName = "n1"
 	pods := []*apiv1.Pod{pod}
 
-	node := BuildTestNode("node", 1000, 1000)
+	node := no.BuildTestNode("node", 1000, 1000)
 
 	resNode, err := SanitizeNode(node, "test-group", taints.TaintConfig{})
 	assert.NoError(t, err)
@@ -42,7 +43,7 @@ func TestSanitizePods(t *testing.T) {
 }
 
 func TestSanitizeLabels(t *testing.T) {
-	oldNode := BuildTestNode("ng1-1", 1000, 1000)
+	oldNode := no.BuildTestNode("ng1-1", 1000, 1000)
 	oldNode.Labels = map[string]string{
 		apiv1.LabelHostname: "abc",
 		"x":                 "y",
@@ -56,7 +57,7 @@ func TestSanitizeLabels(t *testing.T) {
 }
 
 func TestGetNodeResource(t *testing.T) {
-	node := BuildTestNode("n1", 1000, 2*MiB)
+	node := no.BuildTestNode("n1", 1000, 2*MiB)
 
 	cores := getNodeResource(node, apiv1.ResourceCPU)
 	assert.Equal(t, int64(1), cores)
@@ -68,7 +69,7 @@ func TestGetNodeResource(t *testing.T) {
 	assert.Equal(t, int64(0), unknownResourceValue)
 
 	// if we have no resources in capacity we expect getNodeResource to return 0
-	nodeWithMissingCapacity := BuildTestNode("n1", 1000, 2*MiB)
+	nodeWithMissingCapacity := no.BuildTestNode("n1", 1000, 2*MiB)
 	nodeWithMissingCapacity.Status.Capacity = apiv1.ResourceList{}
 
 	cores = getNodeResource(nodeWithMissingCapacity, apiv1.ResourceCPU)
@@ -78,7 +79,7 @@ func TestGetNodeResource(t *testing.T) {
 	assert.Equal(t, int64(0), memory)
 
 	// if we have negative values in resources we expect getNodeResource to return 0
-	nodeWithNegativeCapacity := BuildTestNode("n1", -1000, -2*MiB)
+	nodeWithNegativeCapacity := no.BuildTestNode("n1", -1000, -2*MiB)
 	nodeWithNegativeCapacity.Status.Capacity = apiv1.ResourceList{}
 
 	cores = getNodeResource(nodeWithNegativeCapacity, apiv1.ResourceCPU)
@@ -90,14 +91,14 @@ func TestGetNodeResource(t *testing.T) {
 }
 
 func TestGetNodeCoresAndMemory(t *testing.T) {
-	node := BuildTestNode("n1", 2000, 2048*MiB)
+	node := no.BuildTestNode("n1", 2000, 2048*MiB)
 
 	cores, memory := GetNodeCoresAndMemory(node)
 	assert.Equal(t, int64(2), cores)
 	assert.Equal(t, int64(2048*MiB), memory)
 
 	// if we have no cpu/memory defined in capacity we expect getNodeCoresAndMemory to return 0s
-	nodeWithMissingCapacity := BuildTestNode("n1", 1000, 2*MiB)
+	nodeWithMissingCapacity := no.BuildTestNode("n1", 1000, 2*MiB)
 	nodeWithMissingCapacity.Status.Capacity = apiv1.ResourceList{}
 
 	cores, memory = GetNodeCoresAndMemory(nodeWithMissingCapacity)
@@ -106,11 +107,11 @@ func TestGetNodeCoresAndMemory(t *testing.T) {
 }
 
 func TestGetOldestPod(t *testing.T) {
-	p1 := BuildTestPod("p1", 500, 1000)
+	p1 := po.BuildTestPod("p1", 500, 1000)
 	p1.CreationTimestamp = metav1.NewTime(time.Now().Add(-1 * time.Minute))
-	p2 := BuildTestPod("p2", 500, 1000)
+	p2 := po.BuildTestPod("p2", 500, 1000)
 	p2.CreationTimestamp = metav1.NewTime(time.Now().Add(+1 * time.Minute))
-	p3 := BuildTestPod("p3", 500, 1000)
+	p3 := po.BuildTestPod("p3", 500, 1000)
 	p3.CreationTimestamp = metav1.NewTime(time.Now())
 
 	assert.Equal(t, p1.CreationTimestamp.Time, GetOldestCreateTime([]*apiv1.Pod{p1, p2, p3}))
