+++ cluster-autoscaler/utils/test/test_utils.go
@@ -21,182 +21,12 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"strings"
-	"time"
 
 	"github.com/stretchr/testify/mock"
 	apiv1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/autoscaler/cluster-autoscaler/cloudprovider"
-	kube_types "k8s.io/kubernetes/pkg/kubelet/types"
 )
-
-// BuildTestPod creates a pod with specified resources.
-// [DEPRECATED]: use `NewTestPod` instead
-func BuildTestPod(name string, cpu int64, mem int64) *apiv1.Pod {
-	return NewTestPod(name, WithCPU(cpu), WithMemory(mem))
-}
-
-func basePod(name string) *apiv1.Pod {
-	startTime := metav1.Unix(0, 0)
-	pod := &apiv1.Pod{
-		ObjectMeta: metav1.ObjectMeta{
-			UID:         types.UID(name),
-			Namespace:   "default",
-			Name:        name,
-			SelfLink:    fmt.Sprintf("/api/v1/namespaces/default/pods/%s", name),
-			Annotations: map[string]string{},
-		},
-		Spec: apiv1.PodSpec{
-			Containers: []apiv1.Container{
-				{
-					Resources: apiv1.ResourceRequirements{
-						Requests: apiv1.ResourceList{},
-					},
-				},
-			},
-		},
-		Status: apiv1.PodStatus{
-			StartTime: &startTime,
-		},
-	}
-
-	return pod
-}
-
-// WithCPU adds requests.cpu (unit: number of CPUs) to all containers in the pod
-func WithCPU(cpu int64) func(*apiv1.Pod) {
-	return func(pod *apiv1.Pod) {
-
-		for i := range pod.Spec.Containers {
-			pod.Spec.Containers[i].Resources.Requests[apiv1.ResourceCPU] = *resource.NewQuantity(cpu, resource.DecimalSI)
-		}
-	}
-}
-
-// WithMilliCPU adds requests.cpu (unit: number of milli CPUs) to all containers in the pod
-func WithMilliCPU(cpu int64) func(*apiv1.Pod) {
-	return func(pod *apiv1.Pod) {
-
-		for i := range pod.Spec.Containers {
-			pod.Spec.Containers[i].Resources.Requests[apiv1.ResourceCPU] = *resource.NewMilliQuantity(cpu, resource.DecimalSI)
-		}
-	}
-}
-
-// WithMemory adds requests.memory to all the containers in the pod
-func WithMemory(mem int64) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-
-		for i := range pod.Spec.Containers {
-			pod.Spec.Containers[i].Resources.Requests[apiv1.ResourceMemory] = *resource.NewQuantity(mem, resource.DecimalSI)
-		}
-	}
-}
-
-// WithGPU adds requests.nvidia.com/gpu and limits.nvidia.com/gpu
-// to all the containers in the pod
-func WithGPU(gpusCount int64) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-
-		for i := range pod.Spec.Containers {
-			pod.Spec.Containers[i].Resources.Requests[resourceNvidiaGPU] = *resource.NewQuantity(gpusCount, resource.DecimalSI)
-			pod.Spec.Containers[i].Resources.Limits[resourceNvidiaGPU] = *resource.NewQuantity(gpusCount, resource.DecimalSI)
-		}
-	}
-}
-
-// WithGPUToleration adds nvidia.com/gpu:Exists toleration to the pod
-func WithGPUToleration() func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.Spec.Tolerations = append(pod.Spec.Tolerations, apiv1.Toleration{Key: resourceNvidiaGPU, Operator: apiv1.TolerationOpExists})
-	}
-}
-
-// WithAnnotations adds annotations to the pods
-// note: if a key already exists, it will be overwritten
-func WithAnnotations(anns map[string]string) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-
-		for k, v := range anns {
-			pod.Annotations[k] = v
-		}
-	}
-}
-
-// WithOwnerRef adds owner ref to pod
-func WithOwnerRef(ownerRefs []metav1.OwnerReference) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.OwnerReferences = append(pod.OwnerReferences, ownerRefs...)
-	}
-}
-
-// WithEphemeralStorage adds requests.ephemeral-storage to all the containers in the pod
-func WithEphemeralStorage(ephemeralStorage int64) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		for i := range pod.Spec.Containers {
-			pod.Spec.Containers[i].Resources.Requests[apiv1.ResourceEphemeralStorage] = *resource.NewQuantity(ephemeralStorage, resource.DecimalSI)
-		}
-	}
-}
-
-// ScheduledOnNode sets pod's spec.nodeName
-func ScheduledOnNode(nodeName string) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.Spec.NodeName = nodeName
-	}
-}
-
-// WithStaticPodAnnotation adds 'kubernetes.io/config.source: file' annotation to the pod
-func WithStaticPodAnnotation() func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.Annotations[kube_types.ConfigSourceAnnotationKey] = kube_types.FileSource
-	}
-}
-
-// AsDaemonSetPod adds daemonset owner ref to the pod
-func AsDaemonSetPod(uid string) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.OwnerReferences = GenerateOwnerReferences("ds", "DaemonSet", "apps/v1", types.UID(uid))
-	}
-}
-
-// AsReplicaSetPod adds replicaset owner ref to the pod
-func AsReplicaSetPod(uid string, rsName string) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.OwnerReferences = GenerateOwnerReferences(rsName, "ReplicaSet", "extensions/v1beta1", types.UID(uid))
-	}
-}
-
-// WithMirrorPodAnnotation adds 'kubernetes.io/config.mirror: mirror' annotation to the pod
-func WithMirrorPodAnnotation(uid string) func(*apiv1.Pod) {
-
-	return func(pod *apiv1.Pod) {
-		pod.ObjectMeta.Annotations[kube_types.ConfigMirrorAnnotationKey] = "mirror"
-	}
-}
-
-// NewTestPod creates pod with options passed as functions
-func NewTestPod(name string, options ...func(*apiv1.Pod)) *apiv1.Pod {
-	pod := basePod(name)
-
-	for _, o := range options {
-		o(pod)
-	}
-
-	return pod
-}
 
 // BuildServiceTokenProjectedVolumeSource returns a ProjectedVolumeSource with SA token
 // projection
@@ -214,143 +44,10 @@ func BuildServiceTokenProjectedVolumeSource(path string) *apiv1.ProjectedVolumeS
 
 const (
 	// cannot use constants from gpu module due to cyclic package import
-	resourceNvidiaGPU = "nvidia.com/gpu"
-	gpuLabel          = "cloud.google.com/gke-accelerator"
-	defaultGPUType    = "nvidia-tesla-k80"
+	ResourceNvidiaGPU = "nvidia.com/gpu"
+	GPULabel          = "cloud.google.com/gke-accelerator"
+	DefaultGPUType    = "nvidia-tesla-k80"
 )
-
-// BuildTestNode creates a node with specified capacity.
-func BuildTestNode(name string, millicpu int64, mem int64) *apiv1.Node {
-	node := baseNode(name)
-
-	if millicpu >= 0 {
-		node.Status.Capacity[apiv1.ResourceCPU] = *resource.NewMilliQuantity(millicpu, resource.DecimalSI)
-	}
-	if mem >= 0 {
-		node.Status.Capacity[apiv1.ResourceMemory] = *resource.NewQuantity(mem, resource.DecimalSI)
-	}
-
-	node.Status.Allocatable = apiv1.ResourceList{}
-	for k, v := range node.Status.Capacity {
-		node.Status.Allocatable[k] = v
-	}
-
-	return node
-}
-
-func baseNode(name string) *apiv1.Node {
-	return &apiv1.Node{
-		ObjectMeta: metav1.ObjectMeta{
-			Name:     name,
-			SelfLink: fmt.Sprintf("/api/v1/nodes/%s", name),
-			Labels:   map[string]string{},
-		},
-		Spec: apiv1.NodeSpec{
-			ProviderID: name,
-		},
-		Status: apiv1.NodeStatus{
-			Capacity: apiv1.ResourceList{
-				apiv1.ResourcePods: *resource.NewQuantity(100, resource.DecimalSI),
-			},
-		},
-	}
-}
-
-// AddEphemeralStorageToNode adds ephemeral storage capacity to a given node.
-func AddEphemeralStorageToNode(node *apiv1.Node, eph int64) *apiv1.Node {
-	if eph >= 0 {
-		node.Status.Capacity[apiv1.ResourceEphemeralStorage] = *resource.NewQuantity(eph, resource.DecimalSI)
-		node.Status.Allocatable[apiv1.ResourceEphemeralStorage] = *resource.NewQuantity(eph, resource.DecimalSI)
-	}
-	return node
-}
-
-// AddGpusToNode adds GPU capacity to given node. Default accelerator type is used.
-func AddGpusToNode(node *apiv1.Node, gpusCount int64) {
-	node.Spec.Taints = append(
-		node.Spec.Taints,
-		apiv1.Taint{
-			Key:    resourceNvidiaGPU,
-			Value:  "present",
-			Effect: "NoSchedule",
-		})
-	node.Status.Capacity[resourceNvidiaGPU] = *resource.NewQuantity(gpusCount, resource.DecimalSI)
-	node.Status.Allocatable[resourceNvidiaGPU] = *resource.NewQuantity(gpusCount, resource.DecimalSI)
-	AddGpuLabelToNode(node)
-}
-
-// AddGpuLabelToNode adds GPULabel to give node. This is used to mock intermediate result that GPU on node is not ready
-func AddGpuLabelToNode(node *apiv1.Node) {
-	node.Labels[gpuLabel] = defaultGPUType
-}
-
-// GetGPULabel return GPULabel on the node. This is only used in unit tests.
-func GetGPULabel() string {
-	return gpuLabel
-}
-
-// GetGpuConfigFromNode returns the GPU of the node if it has one. This is only used in unit tests.
-func GetGpuConfigFromNode(node *apiv1.Node) *cloudprovider.GpuConfig {
-	gpuType, hasGpuLabel := node.Labels[gpuLabel]
-	gpuAllocatable, hasGpuAllocatable := node.Status.Allocatable[resourceNvidiaGPU]
-	if hasGpuLabel || (hasGpuAllocatable && !gpuAllocatable.IsZero()) {
-		return &cloudprovider.GpuConfig{
-			Label:        gpuLabel,
-			Type:         gpuType,
-			ResourceName: resourceNvidiaGPU,
-		}
-	}
-	return nil
-}
-
-// SetNodeReadyState sets node ready state to either ConditionTrue or ConditionFalse.
-func SetNodeReadyState(node *apiv1.Node, ready bool, lastTransition time.Time) {
-	if ready {
-		SetNodeCondition(node, apiv1.NodeReady, apiv1.ConditionTrue, lastTransition)
-	} else {
-		SetNodeCondition(node, apiv1.NodeReady, apiv1.ConditionFalse, lastTransition)
-		node.Spec.Taints = append(node.Spec.Taints, apiv1.Taint{
-			Key:    "node.kubernetes.io/not-ready",
-			Value:  "true",
-			Effect: apiv1.TaintEffectNoSchedule,
-		})
-	}
-}
-
-// SetNodeNotReadyTaint sets the not ready taint on node.
-func SetNodeNotReadyTaint(node *apiv1.Node) {
-	node.Spec.Taints = append(node.Spec.Taints, apiv1.Taint{Key: apiv1.TaintNodeNotReady, Effect: apiv1.TaintEffectNoSchedule})
-}
-
-// RemoveNodeNotReadyTaint removes the not ready taint.
-func RemoveNodeNotReadyTaint(node *apiv1.Node) {
-	var final []apiv1.Taint
-	for i := range node.Spec.Taints {
-		if node.Spec.Taints[i].Key == apiv1.TaintNodeNotReady {
-			continue
-		}
-		final = append(final, node.Spec.Taints[i])
-	}
-	node.Spec.Taints = final
-}
-
-// SetNodeCondition sets node condition.
-func SetNodeCondition(node *apiv1.Node, conditionType apiv1.NodeConditionType, status apiv1.ConditionStatus, lastTransition time.Time) {
-	for i := range node.Status.Conditions {
-		if node.Status.Conditions[i].Type == conditionType {
-			node.Status.Conditions[i].LastTransitionTime = metav1.Time{Time: lastTransition}
-			node.Status.Conditions[i].Status = status
-			return
-		}
-	}
-	// Condition doesn't exist yet.
-	condition := apiv1.NodeCondition{
-		Type:               conditionType,
-		Status:             status,
-		LastTransitionTime: metav1.Time{Time: lastTransition},
-	}
-	node.Status.Conditions = append(node.Status.Conditions, condition)
-}
 
 // GenerateOwnerReferences builds OwnerReferences with a single reference
 func GenerateOwnerReferences(name, kind, api string, uid types.UID) []metav1.OwnerReference {
